% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_messages.R
\name{add_ts}
\alias{add_ts}
\title{Add a timestamp to a string}
\usage{
add_ts(...)
}
\arguments{
\item{...}{Zero or more objects which can be coerced to character (and which are pasted together with no separator)}
}
\value{
A timestamped character vector.
}
\description{
Add a timestamp to a string
}
\examples{
# Add a timestamp to a string
add_ts("Some text")

# Multiple objects are concatenated without a space
add_ts("Concatenated", "text")
}
