% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subelement.R
\name{extract_subelement}
\alias{extract_subelement}
\title{Extract subelement from list}
\usage{
extract_subelement(x, n)
}
\arguments{
\item{x}{A list.}

\item{n}{A numeric vector containing the indices of the element to extract.
Must be of length 1 or the same length as x.}
}
\value{
A named vector containing the nth elements of each item in x.
}
\description{
Extract the nth subelement from each item a list
}
\examples{
a_list <- list("foo" = c(1, 2, 3), "bar" = c("a", "b", "c"), "baz" = c(1:10))
print(a_list)

# Get the first element of each item
extract_subelement(a_list, 1)

# Get the nth element
extract_subelement(a_list, c(1, 2, 3))
}
