% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelliott-package.R
\docType{package}
\name{nelliott-package}
\alias{nelliott}
\alias{nelliott-package}
\title{nelliott: Personal library of useful and often-used functions}
\description{
Provides an easy way to access, share, and maintain useful functions that I use across many projects.
}
\author{
\strong{Maintainer}: Nathan Elliott \email{nathan.k.elliott@gmail.com} (\href{https://orcid.org/0000-0002-6005-5308}{ORCID})

}
\keyword{internal}
