% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_messages.R
\name{print_ts}
\alias{print_ts}
\title{Print a timestamped message}
\usage{
print_ts(...)
}
\arguments{
\item{...}{Zero or more objects which can be coerced to character
(and which are pasted together with no separator)}
}
\value{
A timestamped character vector.
}
\description{
A wrapper for the base 'message' function that adds a timestamp to the beginning of the message.
}
\examples{
# Add a timestamp to a printed message
print_ts("Some text")

# Multiple objects are concatenated without a space
print_ts("Concatenated", "text")
}
